Kinda fix bug with caching not working

+++ apps/web/utils/getSchema.ts
@@ -39,11 +39,14 @@ const updateSchemaCache = async (code: string): Promise<ISchema | undefined> => 
     const supabase = createClient(supabaseUrl, supabaseAnonKey)
     const calendar = await ical.fromURL(URLBase + code)
     const events = Object.values(calendar).filter((event:any) => event.type === 'VEVENT').map((event) => getEventFromIcalEvent(event,code))
+    if(events.length ==0){
+        throw new Error("No new events in caching..");
+    }
     await supabase.from('events').delete().match({ parent_calendar: code })
     await supabase.from('events').insert(events)
     await supabase.from('calendars').update({ last_cache: dayjs().format() }).match({ code: code })
     return {events:events,name:code}
- 
+        
  
 }
